Create script to perform translation.
This script can take in a file or a directory (eg. `samples`). If the
flag `-k` / `--keep` is used, the script will attempt to keep memory
errors when performing the translation.

This also requires a `.env` file on your system with `OPENAI_API_KEY`
being set (check the `.env.example` file for an example).